DinhThe fix image user

+++ src/main/webapp/WEB-INF/views/voucher/voucher.jsp
@@ -34,7 +34,7 @@
                     <ul class="navbar-nav me-auto mb-2 mb-lg-0">
                     </ul>
                     <form class="d-flex">
-                        <img src="images.jpg" class="rounded-circle" alt="..." width="40px" height="40px">
+                        <img src="${pageContext.request.contextPath}/images/user.jpg" class="rounded-circle" alt="..." width="40px" height="40px">
                         <div class="dropdown">
                             <button class="btn btn-outline-dark dropdown-toggle ms-2" type="button"
                                     id="dropdownMenuButton1" data-bs-toggle="dropdown" aria-expanded="false">
@@ -146,7 +146,7 @@
                                     data-placement="top" href="/t-shirt-luxury/admin/voucher/getOne?id=${voucher.id}"
                                     title="Chỉnh Sửa"><i class="fa-solid fa-pen-to-square"></i></a>
                             <a class="btn btn-danger rounded-pill" data-toggle="tooltip" data-placement="top"
-                                    title="Xóa" href="/t-shirt-luxury/admin/voucher/delete?id=${voucher.id}"><i class="fa-solid fa-trash"></i></a>
+                                    title="Xóa" href="/t-shirt-luxury/admin/voucher/delete?id=${voucher.id}" onclick="return confirmDelete()"><i class="fa-solid fa-trash"></i></a>
 
                         </td>
                     </tr>
@@ -223,7 +223,10 @@
         </div>
     </div>
 </div>
-
 </body>
-
+<script>
+    confirmDelete = () => {
+        return confirm("Bạn có chắc chắn muốn xóa Voucher này không ?");
+    }
+</script>
 </html>
